-[NIB04.01] Get the list of configurations available for a selected fitting.
Former-commit-id: 0383da8625ade09c6c40801c6156c52737223fbb

+++ NeoCom-Infinity-Backend/src/test/java/org/dimensinfin/eveonline/neocom/infinity/support/miningextraction/converter/CucumberTableToMiningExtractionEntityConverter.java
@@ -31,7 +31,7 @@ public class CucumberTableToMiningExtractionEntityConverter extends CucumberTabl
 				.withOwnerId( Integer.parseInt( cucumberRow.get( OWNER_ID ) ) )
 //				.withDelta( 0L )
 				.withSolarSystemId( Integer.parseInt( cucumberRow.get( SYSTEMID ) ) )
-				.withQuantity( Integer.parseInt( cucumberRow.get( QUANTITY ) ) )
+				.withQuantity( Long.parseLong( cucumberRow.get( QUANTITY ) ) )
 				.withExtractionDateName( cucumberRow.get( EXTRACTIONS_DATE ) )
 				.withExtractionHour( Integer.parseInt( cucumberRow.get( EXTRACTIONS_HOUR ) ) )
 				.build();
